# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Project"
        ]
      },
      "put": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "ProjectController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a project",
        "tags": [
          "Project"
        ]
      },
      "get": {
        "operationId": "ProjectController_search",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Project Name",
            "schema": {
              "example": "Project Name",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (optional) - default 1",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Page size (optional) - default 10",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "techs",
            "required": false,
            "in": "query",
            "description": "Array of technologies",
            "schema": {
              "example": [
                "tech1",
                "tech2"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects found"
          }
        },
        "summary": "Search projects",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/v1/projects/search-per-user": {
      "get": {
        "operationId": "ProjectController_searchPerUser",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search projects per user",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/v1/projects/{projectId}": {
      "get": {
        "operationId": "ProjectController_get",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "example": "ValidUUIDv4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found"
          }
        },
        "summary": "Get a project",
        "tags": [
          "Project"
        ]
      },
      "delete": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "ProjectController_delete",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/v1/projects/{projectId}/reactivate": {
      "patch": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "ProjectController_reactivate",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project reactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivate a project",
        "tags": [
          "Project"
        ]
      }
    }
  },
  "info": {
    "title": "Project Management API",
    "description": "Project Management API",
    "version": "1.0.0",
    "contact": {
      "name": "maulanadityaa",
      "url": "https://github.com/maulanadityaa",
      "email": "maulanadityaaa@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3030"
    },
    {
      "url": "https://project-management-api-cyan.vercel.app"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "test_username",
            "description": "Username"
          },
          "name": {
            "type": "string",
            "example": "test_name",
            "description": "Name"
          },
          "uid": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "example": "example@mail.com",
            "description": "Email"
          }
        },
        "required": [
          "username",
          "name",
          "uid",
          "email"
        ]
      },
...
      "ProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Project Name",
            "description": "Project Name"
          },
          "description": {
            "type": "string",
            "example": "Project Description",
            "description": "Project Description"
          },
          "link": {
            "type": "string",
            "example": "https://project-link.com",
            "description": "Project Link"
          },
          "technologies": {
            "type": "array",
            "example": [
              "tech1",
              "tech2"
            ],
            "description": "Array of technologies",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "example": "image.jpg",
            "description": "Project Image",
            "format": "binary"
          }
        },
        "required": [
          "name",
          "description",
          "link",
          "technologies",
          "image"
        ]
      },
      "ProjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "Project ID"
          },
          "name": {
            "type": "string",
            "example": "Project Name",
            "description": "Project Name"
          },
          "description": {
            "type": "string",
            "example": "Project Description",
            "description": "Project Description"
          },
          "link": {
            "type": "string",
            "example": "https://project-link.com",
            "description": "Project Link"
          },
          "technologies": {
            "example": [
              "tech1",
              "tech2"
            ],
            "description": "Array of technologies",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imageUrl": {
            "type": "string",
            "example": "cloudinary-secure-url",
            "description": "Project Image"
          },
          "userResponse": {
            "description": "User",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01T00:00:00.000Z",
            "description": "Created At"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01T00:00:00.000Z",
            "description": "Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "link",
          "technologies",
          "imageUrl",
          "userResponse",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProjectUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "Project ID"
          },
          "name": {
            "type": "string",
            "example": "Updated Project Name",
            "description": "Project Name (optional)"
          },
          "description": {
            "type": "string",
            "example": "Updated Project Description",
            "description": "Project Description (optional)"
          },
          "link": {
            "type": "string",
            "example": "https://updated-project-link.com",
            "description": "Project Link (optional)"
          },
          "technologies": {
            "type": "array",
            "example": [
              "tech1",
              "tech2"
            ],
            "description": "Array of technologies (optional)",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "example": "image.jpg",
            "description": "Project Image (optional)",
            "format": "binary"
          }
        },
        "required": [
          "id"
        ]
      }
    }
  }
}